Changed include guards to pragma once in gs files. Also fixed some trivial warnings

+++ src/gs/gsCamera.cpp
@@ -42,7 +42,7 @@ Matrix4 gs::Camera::GetProjectionMatrix() const
     return projectionMatrix;
 }
 
-void gs::Camera::Move(const double latitudeChange, const double longitudeChange)
+void gs::Camera::Move(const float latitudeChange, const float longitudeChange)
 {
     latitude += latitudeChange;
     NormalizeLatitude();
